+++ test-data/TestClass.java
@@ -5,5 +5,9 @@ public class TestClass {
 		System.err.println("testing");
 		System.err.println("testing");
 	}
+	
+	public static void main(String[] args) {
+		new TestClass().foo();
+	}
 
 }