+++ CHANGELOG.md
@@ -2,8 +2,7 @@ We use [semantic versioning][semver]
 
 # 2.0.0
 
-- [feature] added `convert` mode to do one-time conversion. The old behaviour is now reachable via the `watch` command.
-  This means that all existing command line invocations of the tool must be changed to include the `watch` command.
+- [feature] added `convert` mode to do one-time conversion. The old behaviour is now reachable via the `watch` command (which is also the default if no command is provided).
 
 # 1.4.1