+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 We use [semantic versioning][semver]
 
+# 2.0.0
+
+- [feature] added `convert` mode to do one-time conversion. The old behaviour is now reachable via the `watch` command (which is also the default if no command is provided).
+
 # 1.4.1
 
 - [tooling] add GitLab build and `dist` task. Update documentation